pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "deepmind-code"
version = "0.1.0"
description = "A powerful, lightweight AI-wrapped CLI for terminal coding."
authors = [
    { name = "Arya Wiratama", email = "aryawiratama2401@gmail.com" },
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "gitpython>=3.1.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
dmc = "deepmind_code.main:cli"

[tool.setuptools.packages.find]
where = ["src"]
